Move attach into home; keep domain as identity
Guest parse and check live next to the form. Domain only routes. Vercel config is no longer named domain. A records are ipv4 again.
