+++ src/domain/host.ts
@@ -1,22 +1,4 @@
-import { hasPlatform } from "./platform";
-import type { Result } from "@/kit/result";
-
-/**
- * Parse a guest-entered domain into a hostname we can claim.
- */
-export function parseDomainHost(input: string): Result<string> {
-  const trimmed = input.trim().toLowerCase();
-  if (trimmed === "") return { ok: false, reason: "Can't use this domain." };
-
-  const withScheme = trimmed.includes("://") ? trimmed : `https://${trimmed}`;
-
-  try {
-    const hostname = new URL(withScheme).hostname;
-    if (hasPlatform(hostname)) {
-      return { ok: false, reason: "Can't use this domain." };
-    }
-    return { ok: true, value: hostname };
-  } catch {
-    return { ok: false, reason: "Can't use this domain." };
-  }
+/** Hostname from a Host header. */
+export function getDomainHost(header: string): string {
+  return header.split(":").at(0)?.toLowerCase() ?? "";
 }
