+++ src/home/check.ts
@@ -1,19 +1,19 @@
+import { getDomainCustom } from "@/domain/custom";
 import type { Result } from "@/kit/result";
-import { getVercelDomainConfig } from "@/vercel/domain";
+import { getVercelDomainConfig } from "@/vercel/config";
 import {
   addVercelProjectDomain,
   getVercelProjectDomain,
   listVercelProjectDomains,
   verifyVercelProjectDomain,
 } from "@/vercel/project";
-import { getDomainCustom } from "./custom";
 import { parseDomainHost } from "./host";
 
 export type DomainCheck =
   | {
       ok: true;
       target: string;
-      config: { cname: Result<string>; apex: Result<string> };
+      config: { cname: Result<string>; ipv4: Result<string> };
       verify: { txt: Result<{ domain: string; value: string }> };
     }
   | { ok: false; reason: string };
@@ -49,7 +49,7 @@ export async function checkDomain(input: string): Promise<DomainCheck> {
     throw new Error("Domain is not on the project");
   }
   const dns = await getVercelDomainConfig({ name: host.value });
-  const config = { cname: dns.cname, apex: dns.ipv4 };
+  const config = { cname: dns.cname, ipv4: dns.ipv4 };
   if (project.value.verified) {
     const txt = { ok: false as const, reason: "Already verified" };
     return { ok: true, target, config, verify: { txt } };
